Switch to using `either` crate
