+++ provider/datagen/Cargo.toml
@@ -72,6 +72,7 @@ proc-macro2 = {version = "1", optional = true }
 
 # Other external dependencies
 displaydoc = { version = "0.2.3", default-features = false }
+either = { workspace = true }
 elsa = "1.10"
 itertools = "0.10"
 log = "0.4"
@@ -93,7 +94,6 @@ ureq = { version = "2", optional = true }
 clap = { version = "4", optional = true, features = ["derive"] }
 eyre = { version = "0.6", optional = true }
 simple_logger = { version = "4.1.0", default-features = false, optional = true }
-either = "1.9.0"
 
 [dev-dependencies]
 crlify = { path = "../../utils/crlify" }
