+++ utils/writeable/Cargo.toml
@@ -20,6 +20,9 @@ independent = true
 [package.metadata.docs.rs]
 all-features = true
 
+[dependencies]
+either = { workspace = true, optional = true }
+
 [dev-dependencies]
 icu_benchmark_macros = { path = "../../tools/benchmark/macros" }
 rand = { version = "0.8", features = ["small_rng"] }
@@ -29,6 +32,7 @@ criterion = { workspace = true }
 
 [features]
 bench = []
+either = ["dep:either"]
 
 [package.metadata.cargo-all-features]
 # Bench feature gets tested separately and is only relevant for CI
