+++ utils/writeable/src/lib.rs
@@ -67,6 +67,8 @@
 extern crate alloc;
 
 mod cmp;
+#[cfg(feature = "either")]
+mod either;
 mod impls;
 mod ops;
 
@@ -316,57 +318,6 @@ pub trait Writeable {
         #[allow(clippy::unwrap_used)] // infallible impl
         self.write_to(&mut wc).unwrap();
         wc.finish().reverse()
-    }
-}
-
-/// A [`Writeable`] that is either one type or another type.
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[allow(clippy::exhaustive_enums)] // defined to be an enumeration over two types
-pub enum EitherWriteable<W0, W1> {
-    /// A value of type `W0`.
-    A(W0),
-    /// A value of type `W1`.
-    B(W1),
-}
-
-impl<W0, W1> Writeable for EitherWriteable<W0, W1>
-where
-    W0: Writeable,
-    W1: Writeable,
-{
-    fn write_to<W: fmt::Write + ?Sized>(&self, sink: &mut W) -> fmt::Result {
-        match self {
-            EitherWriteable::A(w) => w.write_to(sink),
-            EitherWriteable::B(w) => w.write_to(sink),
-        }
-    }
-
-    fn write_to_parts<S: PartsWrite + ?Sized>(&self, sink: &mut S) -> fmt::Result {
-        match self {
-            EitherWriteable::A(w) => w.write_to_parts(sink),
-            EitherWriteable::B(w) => w.write_to_parts(sink),
-        }
-    }
-
-    fn writeable_length_hint(&self) -> LengthHint {
-        match self {
-            EitherWriteable::A(w) => w.writeable_length_hint(),
-            EitherWriteable::B(w) => w.writeable_length_hint(),
-        }
-    }
-
-    fn write_to_string(&self) -> Cow<str> {
-        match self {
-            EitherWriteable::A(w) => w.write_to_string(),
-            EitherWriteable::B(w) => w.write_to_string(),
-        }
-    }
-
-    fn write_cmp_bytes(&self, other: &[u8]) -> core::cmp::Ordering {
-        match self {
-            EitherWriteable::A(w) => w.write_cmp_bytes(other),
-            EitherWriteable::B(w) => w.write_cmp_bytes(other),
-        }
     }
 }
